backend/requirements.txt
Flask==3.0.3
requests==2.32.3
pillow
mistralai
Werkzeug>=3.0.0
flask-cors